Hacky fix for EMI rendering
Fix #2735 by doing a partial revert of the changes to GTRecipeEMICategory in 039659f
I suspect this works due to some implementation detail of EmiTexture/EmiDrawContext setting z to 0 or something?
I couldn't figure out the full details myself; please reject this if you know the "proper" way to fix this.